feat: refactor user module structure and update user information handling
