+++ backend/src/modules/auth/auth.module.ts
@@ -2,8 +2,9 @@ import { Module } from '@nestjs/common';
 import { CqrsModule } from '@nestjs/cqrs';
 import { JwtModule } from '@nestjs/jwt/dist/jwt.module';
 
+import { UserRepository } from 'src/modules/users/users.repository';
+
 import { DatabaseModule } from '../database/database.module';
-import { UserRepository } from '../user/user.repository';
 
 import { AuthController } from './auth.controller';
 import { AuthRepository } from './auth.repository';
