--- backend/src/modules/user/decorators/index.ts
@@ -1,1 +1,0 @@
-export * from './myInformation.decorator';
